requirements.txt
asgiref==3.8.1
Django==4.2.6
django-cors-headers==4.7.0
django-filter==2.4.0
djangorestframework==3.15.2
sqlparse==0.5.3
